Fixed various errors and made it look pretty
Fixed an error that would caue the report page to repeat the same
values. Added a banner and moved the user controls to the top. Cut about
a hundred lines of code form the setXML function to make the logic flow
easier for whoever picks the project up after me. And am currently
working on the logic for adding list of values to attributes.